refactor: rename the endorsement package to policy
The package holds the appraisal-policy language, and in RATS terms its
contents are reference values and appraisal policies, not endorsements
(the hardware-vendor collateral entries are). Split the schema into
artifact.go, sev.go, tdx.go, and shape.go so the platform-specific
blocks are separated from the platform-neutral core.
