+++ verifier/policy/artifact_test.go
@@ -1,4 +1,4 @@
-package endorsement
+package policy
 
 import (
 	"os"
