+++ verifier/policy/testdata/platform-endorsements.json
